Reader and adapter implementation

+++ src/main/java/com/bmc/truesight/remedy/beans/FieldItem.java
@@ -4,29 +4,29 @@ import java.util.Map;
 
 public class FieldItem {
 
-	private long fieldId;
-	Map<String, String> valueMap;
+    private Integer fieldId;
+    Map<String, String> valueMap;
 
-	public Map<String, String> getValueMap() {
-		return valueMap;
-	}
+    public Map<String, String> getValueMap() {
+        return valueMap;
+    }
 
-	public void setValueMap(Map<String, String> valueMap) {
-		this.valueMap = valueMap;
-	}
+    public void setValueMap(Map<String, String> valueMap) {
+        this.valueMap = valueMap;
+    }
 
-	public long getFieldId() {
-		return fieldId;
-	}
+    public Integer getFieldId() {
+        return fieldId;
+    }
 
-	public void setFieldId(long fieldId) {
-		this.fieldId = fieldId;
-	}
+    public void setFieldId(Integer fieldId) {
+        this.fieldId = fieldId;
+    }
 
-	@Override
-	public String toString() {
-		return "FieldItem [ fieldId=" + fieldId + ", valueMap=" + valueMap
-				+ "]";
-	}
+    @Override
+    public String toString() {
+        return "FieldItem [ fieldId=" + fieldId + ", valueMap=" + valueMap
+                + "]";
+    }
 
 }
